fixing some JIRAs
git-svn-id: https://svn.codehaus.org/groovy/eclipse/trunk@17154 a5544e8c-8a19-0410-ba12-f9af4593a198

+++ ide-test/org.codehaus.groovy.eclipse.core.test/jdt-src/org/eclipse/jdt/core/tests/util/Util.java
@@ -1335,7 +1335,7 @@ public static String getProblemLog(CompilationResult compilationResult, boolean 
 					if (showWarningToken) {
 						long irritant = ProblemReporter.getIrritant(problem.getID());
 						if (irritant != 0) {
-							String warningToken = CompilerOptions.warningTokenFromIrritant(irritant);
+							String warningToken = CompilerOptions.warningTokenFromIrritant((int) irritant);
 							if (warningToken != null) {
 								buffer.append("[@sup:").append(warningToken).append("] ");
 							}
